requirements.txt
netmiko>=4.0.0
nornir>=3.0.0
nornir_netmiko>=3.0.0
nornir_utils>=3.0.0
napalm>=4.0.0
pyyaml>=6.0
jinja2>=3.0
pytest>=7.0
